+++ src/index.js
@@ -5,8 +5,8 @@ import { LocaleProvider } from 'antd';
 import enUS from 'antd/lib/locale-provider/en_US';
 
 // Load Global Styles
-import './styles/reset.css';
-import './styles/global.css';
+import './ui/styles/reset.css';
+import './ui/styles/global.css';
 
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
