+++ src/ui/utils/config/spaces.js
