+++ src/ui/utils/mixins/paint.js
